Update group creation and handling
Lights are not required in a group creation call. This removes this
requirement to better work with the official Hue app. This also reworks
group light accounting to operate the right way with a list of light IDs
instead of bitshift storage so that lights with more complicated names
can be added to a group.